Reorganise @brainstaff/rbac sources.
